+++ Magic/src/main/java/com/elmakers/mine/bukkit/integration/MobArenaManager.java
@@ -2,20 +2,32 @@ package com.elmakers.mine.bukkit.integration;
 
 import com.elmakers.mine.bukkit.api.magic.Mage;
 import com.elmakers.mine.bukkit.api.magic.MageController;
+import com.elmakers.mine.bukkit.protection.BlockBreakManager;
+import com.elmakers.mine.bukkit.protection.BlockBuildManager;
+import com.garbagemule.MobArena.MobArena;
 import com.garbagemule.MobArena.events.ArenaPlayerJoinEvent;
 import com.garbagemule.MobArena.events.ArenaPlayerLeaveEvent;
+import com.garbagemule.MobArena.framework.Arena;
+import com.garbagemule.MobArena.framework.ArenaMaster;
 import org.bukkit.Bukkit;
+import org.bukkit.block.Block;
+import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.plugin.Plugin;
 
+import java.util.List;
 import java.util.Set;
 
-public class MobArenaManager implements Listener {
+public class MobArenaManager implements Listener, BlockBreakManager, BlockBuildManager {
     private final MageController controller;
+    private boolean protect;
+    private List<String> protectedArenas = null;
+    private MobArena mobArena;
 
-    public MobArenaManager(MageController controller) {
+    public MobArenaManager(MageController controller, Plugin plugin, ConfigurationSection configuration) {
         this.controller = controller;
 
         Set<String> magicMobKeys = controller.getMobKeys();
@@ -24,8 +36,25 @@ public class MobArenaManager implements Listener {
             String mobKey = mob.toLowerCase().replaceAll("[-_\\.]", "");
             new MagicMACreature(controller, mobKey, controller.getMob(mob));
         }
+        
+        if (plugin instanceof MobArena) {
+            mobArena = (MobArena)plugin;
+        }
+        configure(configuration);
 
         Bukkit.getPluginManager().registerEvents(this, controller.getPlugin());
+    }
+    
+    public void configure(ConfigurationSection configuration) {
+        protect = configuration.getBoolean("protect");
+        if (protect) {
+            controller.getLogger().info("Spells that break or build blocks can't be used in MobArenas");
+        }
+        if (configuration.contains("protected")) {
+            protectedArenas = configuration.getStringList("protected");
+        } else {
+            protectedArenas = null;
+        }
     }
 
     @EventHandler
@@ -51,5 +80,35 @@ public class MobArenaManager implements Listener {
         if (!s.startsWith("magic:")) return null;
         s = s.substring(6);
         return controller.createItem(s);
+    }
+    
+    public boolean isProtected() {
+        return mobArena != null && (protect || protectedArenas != null);
+    }
+
+    @Override
+    public boolean hasBreakPermission(Player player, Block block) {
+        ArenaMaster am = mobArena.getArenaMaster();
+        if (protect) {
+            for (Arena arena : am.getArenas()) {
+                if (arena.isProtected() && arena.getRegion().contains(block.getLocation())) {
+                    return false;
+                }
+            }
+        }
+        if (protectedArenas != null) {
+            for (String arenaName : protectedArenas) {
+                Arena arena = am.getArenaWithName(arenaName);
+                if (arena != null && arena.isProtected() && arena.getRegion().contains(block.getLocation())) {
+                    return false;
+                }
+            }
+        }
+        return true;
+    }
+
+    @Override
+    public boolean hasBuildPermission(Player player, Block block) {
+        return hasBreakPermission(player, block);
     }
 }